feat(refactor): refactor group bbox and tag to use new implementation instead of from draw functions
