+++ peekingduck/pipeline/nodes/draw/group_bbox_and_tag.py
@@ -20,7 +20,8 @@ from typing import Any, Dict, List
 
 import numpy as np
 from peekingduck.pipeline.nodes.node import AbstractNode
-from peekingduck.pipeline.nodes.draw.utils.drawfunctions import draw_bboxes, draw_tags
+from peekingduck.pipeline.nodes.draw.utils.bbox import draw_bboxes, draw_tags
+from peekingduck.pipeline.nodes.draw.utils.constants import TOMATO
 
 
 class Node(AbstractNode):
@@ -50,10 +51,8 @@ class Node(AbstractNode):
         group_tags = self._get_group_tags(
             inputs["large_groups"], self.tag)
 
-        draw_bboxes(inputs["img"], group_bboxes,
-                    self.bbox_color, self.bbox_thickness)   # type: ignore
-        draw_tags(inputs["img"], group_bboxes,
-                  group_tags, self.bbox_color)  # type: ignore
+        draw_bboxes(inputs["img"], group_bboxes, TOMATO)   # type: ignore
+        draw_tags(inputs["img"], group_bboxes, group_tags, TOMATO)  # type: ignore
 
         return {}
